functions/requirements.txt
azure-functions
azure-cognitiveservices-vision-customvision
azure-storage
numpy
Pillow
pyodbc